+++ client/ducks/loginIdentitySlice.ts
@@ -36,10 +36,12 @@ const loginRequest = createAsyncThunk(
 interface loginIdentity {
   isLogin: boolean;
   userId: number | null;
+  notificationToken: string | null;
 }
 const initialState: loginIdentity = {
   isLogin: false,
   userId: null,
+  notificationToken: null,
 };
 
 export const loginIdentitySlice = createSlice({
@@ -49,11 +51,15 @@ export const loginIdentitySlice = createSlice({
     initLoginIdentity: (state: loginIdentity): void => {
       state.isLogin = false;
       state.userId = null;
+      state.notificationToken = null;
       removeCookie('accessJwtToken');
     },
     oauthLogin: (state, action): void => {
       state.isLogin = true;
       state.userId = action.payload;
+    },
+    notificationToken: (state, action: PayloadAction<string>): void => {
+      state.notificationToken = action.payload;
     },
   },
   extraReducers: (builder): void => {
@@ -66,5 +72,6 @@ export const loginIdentitySlice = createSlice({
     );
   },
 });
-export const { initLoginIdentity, oauthLogin } = loginIdentitySlice.actions;
+export const { initLoginIdentity, oauthLogin, notificationToken } =
+  loginIdentitySlice.actions;
 export { loginRequest };
