+++ client/module/kakaoShare.tsx
@@ -47,12 +47,12 @@ export function KaKaoShare({ habitData }: IhabitData) {
           link: {
             mobileWebUrl:
               habitData === undefined
-                ? 'http://13.209.20.62:3000/'
-                : `http://13.209.20.62:3000//habit/detail/${habitData.habitId}`,
+                ? 'https://66challenge.shop'
+                : `https://66challenge.shop/habit/detail/${habitData.habitId}`,
             webUrl:
               habitData === undefined
-                ? 'http://13.209.20.62:3000/'
-                : `http://13.209.20.62:3000//habit/detail/${habitData.habitId}`,
+                ? 'https://66challenge.shop'
+                : `https://66challenge.shop/habit/detail/${habitData.habitId}`,
           },
         },
         buttons: [
@@ -61,12 +61,12 @@ export function KaKaoShare({ habitData }: IhabitData) {
             link: {
               mobileWebUrl:
                 habitData === undefined
-                  ? 'http://13.209.20.62:3000/'
-                  : `http://13.209.20.62:3000//habit/detail/${habitData.habitId}`,
+                  ? 'https://66challenge.shop'
+                  : `https://66challenge.shop/habit/detail/${habitData.habitId}`,
               webUrl:
                 habitData === undefined
-                  ? 'http://13.209.20.62:3000/'
-                  : `http://13.209.20.62:3000//habit/detail/${habitData.habitId}`,
+                  ? 'https://66challenge.shop'
+                  : `https://66challenge.shop/habit/detail/${habitData.habitId}`,
             },
           },
         ],
