+++ client/pages/_app.tsx
@@ -6,18 +6,7 @@ import type { NextPage } from 'next';
 import type { AppProps } from 'next/app';
 import Layout from '../components/layout/layout';
 import { LoadingIndicator } from '../components/loadingIndicator';
-import firebase from 'firebase';
-import { useEffect } from 'react';
-import { getToken } from '../util/firebase';
-const firebaseConfig = {
-  apiKey: process.env.NEXT_PUBLIC_FIREBASE_API_KEY,
-  authDomain: process.env.NEXT_PUBLIC_FIREBASE_AUTH_DOMAIN,
-  projectId: process.env.NEXT_PUBLIC_FIREBASE_PROJECT_ID,
-  storageBucket: process.env.NEXT_PUBLIC_FIREBASE_STORAGE_BUCKET,
-  messagingSenderId: process.env.NEXT_PUBLIC_FIREBASE_MESSAGING_SENDER_ID,
-  appId: process.env.NEXT_PUBLIC_FIREBASE_APP_ID,
-  measurementId: process.env.NEXT_PUBLIC_FIREBASE_MEASUREMENT_ID,
-};
+
 export type NextPageWithLayout<P = {}, IP = P> = NextPage<P, IP> & {
   getLayout?: (page: ReactElement) => ReactNode;
 };
@@ -27,18 +16,7 @@ type AppPropsWithLayout = AppProps & {
 };
 function MyApp({ Component, pageProps }: AppPropsWithLayout) {
   const getLayout = Component.getLayout || ((page) => page);
-  if (!firebase.apps.length) {
-    firebase.initializeApp(firebaseConfig);
-  } else {
-    firebase.app();
-  }
-  useEffect(() => {
-    firebaseMessageToken();
-  }, []);
-  // async 를 사용학 위해 메서드로 따로 분리함
-  const firebaseMessageToken = async () => {
-    let token = await getToken();
-  };
+
   return (
     <PersistGate persistor={persistor} loading={<LoadingIndicator />}>
       {Component.getLayout ? (
